+++ src/RcppExports.cpp
@@ -37,15 +37,16 @@ BEGIN_RCPP
 END_RCPP
 }
 // CPL_geodetic_azimuth
-Rcpp::NumericVector CPL_geodetic_azimuth(Rcpp::List sfc, double semi_major, double inv_flattening);
-RcppExport SEXP _lwgeom_CPL_geodetic_azimuth(SEXP sfcSEXP, SEXP semi_majorSEXP, SEXP inv_flatteningSEXP) {
+Rcpp::NumericVector CPL_geodetic_azimuth(Rcpp::List sfc, double semi_major, double inv_flattening, Nullable<Rcpp::List> sfc2_);
+RcppExport SEXP _lwgeom_CPL_geodetic_azimuth(SEXP sfcSEXP, SEXP semi_majorSEXP, SEXP inv_flatteningSEXP, SEXP sfc2_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::List >::type sfc(sfcSEXP);
     Rcpp::traits::input_parameter< double >::type semi_major(semi_majorSEXP);
     Rcpp::traits::input_parameter< double >::type inv_flattening(inv_flatteningSEXP);
-    rcpp_result_gen = Rcpp::wrap(CPL_geodetic_azimuth(sfc, semi_major, inv_flattening));
+    Rcpp::traits::input_parameter< Nullable<Rcpp::List> >::type sfc2_(sfc2_SEXP);
+    rcpp_result_gen = Rcpp::wrap(CPL_geodetic_azimuth(sfc, semi_major, inv_flattening, sfc2_));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -179,6 +180,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::List >::type sfc(sfcSEXP);
     Rcpp::traits::input_parameter< int >::type prec(precSEXP);
     rcpp_result_gen = Rcpp::wrap(CPL_geohash(sfc, prec));
+    return rcpp_result_gen;
+END_RCPP
+}
+// CPL_bbox_from_geohash
+Rcpp::NumericVector CPL_bbox_from_geohash(Rcpp::CharacterVector h, int prec);
+RcppExport SEXP _lwgeom_CPL_bbox_from_geohash(SEXP hSEXP, SEXP precSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type h(hSEXP);
+    Rcpp::traits::input_parameter< int >::type prec(precSEXP);
+    rcpp_result_gen = Rcpp::wrap(CPL_bbox_from_geohash(h, prec));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -349,7 +362,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_lwgeom_CPL_geodetic_area", (DL_FUNC) &_lwgeom_CPL_geodetic_area, 3},
     {"_lwgeom_CPL_geodetic_length", (DL_FUNC) &_lwgeom_CPL_geodetic_length, 3},
-    {"_lwgeom_CPL_geodetic_azimuth", (DL_FUNC) &_lwgeom_CPL_geodetic_azimuth, 3},
+    {"_lwgeom_CPL_geodetic_azimuth", (DL_FUNC) &_lwgeom_CPL_geodetic_azimuth, 4},
     {"_lwgeom_CPL_geodetic_segmentize", (DL_FUNC) &_lwgeom_CPL_geodetic_segmentize, 2},
     {"_lwgeom_CPL_geodetic_covers", (DL_FUNC) &_lwgeom_CPL_geodetic_covers, 2},
     {"_lwgeom_CPL_geodetic_distance", (DL_FUNC) &_lwgeom_CPL_geodetic_distance, 7},
@@ -361,6 +374,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_lwgeom_CPL_split", (DL_FUNC) &_lwgeom_CPL_split, 2},
     {"_lwgeom_CPL_wrap_x", (DL_FUNC) &_lwgeom_CPL_wrap_x, 3},
     {"_lwgeom_CPL_geohash", (DL_FUNC) &_lwgeom_CPL_geohash, 2},
+    {"_lwgeom_CPL_bbox_from_geohash", (DL_FUNC) &_lwgeom_CPL_bbox_from_geohash, 2},
     {"_lwgeom_CPL_lwgeom_transform", (DL_FUNC) &_lwgeom_CPL_lwgeom_transform, 2},
     {"_lwgeom_CPL_minimum_bounding_circle", (DL_FUNC) &_lwgeom_CPL_minimum_bounding_circle, 1},
     {"_lwgeom_CPL_subdivide", (DL_FUNC) &_lwgeom_CPL_subdivide, 2},