+++ NEWS.md
@@ -1,3 +1,11 @@
+# version 0.2-15
+
+* `st_geod_azimuth()` adds feature/pairwise azimuth #97 by @robitalec 
+
+* add `st_geom_from_geohash()`; #37
+
+* fix default for argument `digits` in `st_astext()`
+
 # version 0.2-14
 
 * `st_perimeter()` is deprecated in favor of `st_perimeter_lwgeom()`, as `sf` takes over with `sf::st_perimeter()`.